docs(menu): Consolidate ContextMenu example
